Updated to flash in morse code
Updated code to flash LED in morse code. Removed extraneous files.
