+++ Blink_LED/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "particle.targetDevice": "",
+    "particle.firmwareVersion": "6.1.1"
+}
